# docs/swagger.yaml
definitions:
  dto.CreateSubscription:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateSubscriptionResponse:
    properties:
      id:
        type: integer
    type: object
  dto.DeleteSubscriptionResponse:
    properties:
      deleted:
        type: boolean
    type: object
  dto.GetSubscriptionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.GetSubscriptionResponse'
        type: array
    type: object
  dto.GetSubscriptionResponse:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.GetTotalSumResponse:
    properties:
      total_sum:
        type: integer
    type: object
  dto.UpdateSubscription:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateSubscriptionResponse:
    properties:
      updated:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Returns paginated list of subscriptions
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetSubscriptionListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new user subscription
      parameters:
      - description: Subscription data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Deletes subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Returns a single subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Updates subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Calculates total sum of subscriptions for given filters
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Start date (DD-MM-YYYY)
        in: query
        name: start_date
        type: string
      - description: End date (DD-MM-YYYY)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetTotalSumResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get total subscription cost
      tags:
      - subscriptions
swagger: "2.0"